NPC Sheet fixes

+++ modules/item/item-sheet.js
@@ -82,8 +82,6 @@ export default class ItemSheetWfrp4e extends ItemSheet {
       else {
         data["loreValue"] = this.item.lore.value;
       }
-      data["descriptionAndLore"] = WFRP_Utility._spellDescription(this.item.data)
-
     }
 
 
@@ -196,19 +194,24 @@ export default class ItemSheetWfrp4e extends ItemSheet {
 
   // Lore input is tricky because we need to choose from a set of defined choices, but it isn't a dropdown
   async _onLoreChange(event) {
-    let loreEffects = this.item.data.effects.filter(i => i.flags.wfrp4e.lore)
-    await this.item.deleteEmbeddedEntity("ActiveEffect", loreEffects.map(i => i._id))
-    let inputLore = event.target.value;
-    // Go through each lore name
-    for (let lore in game.wfrp4e.config.magicLores) {
-      // If lore value matches config, use that (Update the actor with the "key" value)
-      if (inputLore == game.wfrp4e.config.magicLores[lore]) {
-        this.item.createEmbeddedDocuments("ActiveEffect", [game.wfrp4e.config.loreEffects[lore]])
-        return this.item.update({ 'data.lore.value': lore });
+    if (!this.item.isOwned)
+    {
+      let loreEffects = this.item.effects.filter(i => i.flags.wfrp4e.lore)
+      await this.item.deleteEmbeddedEntity("ActiveEffect", loreEffects.map(i => i._id))
+      let inputLore = event.target.value;
+      // Go through each lore name
+      for (let lore in game.wfrp4e.config.magicLores) {
+        // If lore value matches config, use that (Update the actor with the "key" value)
+        if (inputLore == game.wfrp4e.config.magicLores[lore]) {
+          this.item.createEmbeddedDocuments("ActiveEffect", [game.wfrp4e.config.loreEffects[lore]])
+          return this.item.update({ 'data.lore.value': lore });
+        }
       }
+      // Otherwise, if the input isn't recognized, store user input directly as a custom lore
+      return this.item.update({ 'data.lore.value': inputLore });
     }
-    // Otherwise, if the input isn't recognized, store user input directly as a custom lore
-    return this.item.update({ 'data.lore.value': inputLore });
+    else 
+      return ui.notifications.error("You cannot edit a spell's lore while it is owned by an actor");
   }
